refactor - image styled components
