+++ src/components/image/index.tsx
@@ -1,55 +1,5 @@
 import React, { useRef } from 'react';
-import styled from 'styled-components';
-import { BsImage } from 'react-icons/bs'; 
-import { PiTrashSimple } from "react-icons/pi";
-
-export const Container = styled.div`
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    gap: 12px;
-`;
-
-export const ImageContainer = styled.div`
-    position: relative;
-    width: 100px;
-    height: 100px;
-    border: 0.5px solid #3B3B3B;
-    border-radius: 40px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    cursor: pointer;
-`;
-
-export const StyledImage = styled.img`
-    display: block;
-    width: 100px;
-    height: 100px;
-    border-radius: 40px;
-`;
-
-const ImageIcon = styled(BsImage)`
-    font-size: 25px;
-    color: #B5B5B5;
-`;
-
-const FileInput = styled.input`
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    opacity: 0;
-    cursor: pointer;
-`;
-
-const CustomDeleteIcon = styled(PiTrashSimple)`
-    color: #A54D27;
-    cursor: pointer;
-    font-size: 2opx;
-     background-color: transparent;
-`;
+import { Container, CustomDeleteIcon, FileInput, ImageContainer, ImageIcon, StyledImage } from './style';
 
 export default function ImageInput({ onDeleteImage, imageUrl, setNewPost }: {onDeleteImage:any, imageUrl:string, setNewPost: any}){
     const inputRef = useRef(null);
@@ -68,8 +18,6 @@ export default function ImageInput({ onDeleteImage, imageUrl, setNewPost }: {onD
             reader.readAsDataURL(file); 
         }
     };
-
-
 
     return (
         <Container>
